base: support cmpxchg with different sizes
Issue #986

+++ base-fiasco/src/base/lock/lock.cc
@@ -38,7 +38,7 @@ void Cancelable_lock::lock()
 	 * XXX: How to notice cancel-blocking signals issued when  being outside the
 	 *      'l4_ipc_sleep' system call?
 	 */
-	while (!Genode::cmpxchg(&_lock, UNLOCKED, LOCKED))
+	while (!Genode::cmpxchg(&_lock, (int)UNLOCKED, (int)LOCKED))
 		if (Fiasco::l4_ipc_sleep(Fiasco::l4_ipc_timeout(0, 0, 500, 0)) != L4_IPC_RETIMEOUT)
 			throw Genode::Blocking_canceled();
 }
